+++ app/Database/Migrations/2021-05-01-001900_Posts.php
@@ -9,36 +9,40 @@ class Posts extends Migration
 	public function up()
 	{
 		$this->forge->addField([
-			'id'                => [
+				'id'                => [
 				'type'           => 'INT',
 				'constraint' => 11,
 				'unsigned'  => true,
 				'auto_increment' => true,
 			],
-			'title'         => [
+				'title'         => [
 				'type'           => 'VARCHAR',
 				'constraint'     => '255',
 			],
-			'slug'     => [
+				'slug'     => [
 				'type'             => 'VARCHAR',
 				'constraint'   => '255',
 			],
-			'body'     => [
+				'body'     => [
 				'type'             => 'LONGTEXT',
 			],			
 				'photo_post'     => [
 				'type'           => 'TEXT',
 				'default'           => null,
+				],			
+				'views'     => [
+				'type'           => 'DECIMAL(10,0)',
+				'default'           => 0,
 			],
-			'created_at'  => [
+				'created_at'  => [
 				'type'	      => 'DATETIME',
 				'null'	      => true,
 			],
-			'updated_at' 	=> [
+				'updated_at' 	=> [
 				'type'	     => 'DATETIME',
 				'null'	     => true,
 			],
-			'deleted_at' 	=> [
+				'deleted_at' 	=> [
 				'type'	     => 'DATETIME',
 				'null'	     => true,
 			]